requirements.txt

langchain==0.2.14
langchain-openai==0.1.22
langchain-community==0.2.12
faiss-cpu==1.8.0.post1
pypdf==4.3.1
python-dotenv==1.0.1
streamlit==1.39.0
tiktoken==0.7.0
